refactor: move CountryEnum back to enums and add JSONSchema annotations to CountryDto
Move CountryEnum from dtos/country.dto back to enums/country.enum and update all imports across DTOs, validators, and tests. Add JSONSchema decorators to CountryDto fields for better OpenAPI documentation. Remove unused GEOGRAPHIC_EUROPEAN_COUNTRIES import from country.functions.ts. Remove unused IsEnum import from client-data.dto.ts.
