+++ dtos/client-data.dto.ts
@@ -1,7 +1,6 @@
 import {
   IsDefined,
   IsEmail,
-  IsEnum,
   IsNotEmpty,
   IsOptional,
   IsPhoneNumber,
