+++ validators/client-data-validator.spec.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata';
 import { validateClientDataDto } from './client-data-validator';
-import { CountryEnum } from '../dtos/country.dto';
+import { CountryEnum } from '../enums/country.enum';
 
 describe('validateClientDataDto', () => {
   it('should return no errors for a valid client data object', async () => {
