+++ validators/company-data-validator.spec.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata';
 import { validateCompanyDataDto } from './company-data-validator';
-import { CountryEnum } from '../dtos/country.dto';
+import { CountryEnum } from '../enums/country.enum';
 import { LanguageEnum } from '../enums/language.enum';
 
 describe('CompanyDataDto Validator', () => {
