+++ validators/product-downgradable-request-validator.spec.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 import { validateProductDowngradableRequestDto } from './product-downgradable-request-validator';
-import { CountryEnum } from '../dtos/country.dto';
 import { DurationEnum } from '../enums/duration.enum';
+import { CountryEnum } from '../enums/country.enum';
 
 describe('validateProductDowngradableRequestDto', () => {
   it('should return no errors for a valid request', async () => {
